Migrate paymentRequestStatusReadable over into utils/formatters
