+++ src/pages/Office/PaymentRequestQueue/PaymentRequestQueue.jsx
@@ -6,8 +6,8 @@ import styles from './PaymentRequestQueue.module.scss';
 import { usePaymentRequestQueueQueries, useUserQueries } from 'hooks/queries';
 import { createHeader } from 'components/Table/utils';
 import { HistoryShape } from 'types/router';
-import { formatAgeToDays, paymentRequestStatusReadable } from 'shared/formatters';
-import { formatDateFromIso, serviceMemberAgencyLabel } from 'utils/formatters';
+import { formatAgeToDays } from 'shared/formatters';
+import { formatDateFromIso, serviceMemberAgencyLabel, paymentRequestStatusReadable } from 'utils/formatters';
 import MultiSelectCheckBoxFilter from 'components/Table/Filters/MultiSelectCheckBoxFilter';
 import SelectFilter from 'components/Table/Filters/SelectFilter';
 import DateSelectFilter from 'components/Table/Filters/DateSelectFilter';
