+++ src/utils/formatters.test.js
@@ -1,5 +1,7 @@
 import * as formatters from './formatters';
 
+import PAYMENT_REQUEST_STATUS from 'constants/paymentRequestStatus';
+
 describe('formatters', () => {
   describe('format date for customer app', () => {
     it('should format customer date to DD MMM YYYY', () => {
@@ -180,6 +182,36 @@ describe('formatters', () => {
       expect(formatters.formatToOrdinal(2)).toEqual('2nd');
       expect(formatters.formatToOrdinal(3)).toEqual('3rd');
       expect(formatters.formatToOrdinal(4)).toEqual('4th');
+    });
+  });
+
+  describe('paymentRequestStatusReadable', () => {
+    it('returns expected string for PENDING', () => {
+      expect(formatters.paymentRequestStatusReadable(PAYMENT_REQUEST_STATUS.PENDING)).toEqual('Payment requested');
+    });
+
+    it('returns expected string for REVIEWED', () => {
+      expect(formatters.paymentRequestStatusReadable(PAYMENT_REQUEST_STATUS.REVIEWED)).toEqual('Reviewed');
+    });
+
+    it('returns expected string for SENT_TO_GEX', () => {
+      expect(formatters.paymentRequestStatusReadable(PAYMENT_REQUEST_STATUS.SENT_TO_GEX)).toEqual('Reviewed');
+    });
+
+    it('returns expected string for RECEIVED_BY_GEX', () => {
+      expect(formatters.paymentRequestStatusReadable(PAYMENT_REQUEST_STATUS.RECEIVED_BY_GEX)).toEqual('Reviewed');
+    });
+
+    it('returns expected string for PAID', () => {
+      expect(formatters.paymentRequestStatusReadable(PAYMENT_REQUEST_STATUS.PAID)).toEqual('Paid');
+    });
+
+    it('returns expected string for EDI_ERROR', () => {
+      expect(formatters.paymentRequestStatusReadable(PAYMENT_REQUEST_STATUS.EDI_ERROR)).toEqual('EDI error');
+    });
+
+    it('returns expected string for DEPRECATED', () => {
+      expect(formatters.paymentRequestStatusReadable(PAYMENT_REQUEST_STATUS.DEPRECATED)).toEqual('Deprecated');
     });
   });
 });
